requirements.txt
Django==2.2.24
djangorestframework==3.13.1
drf-writable-nested==0.6.4
drf-yasg==1.21.10
exceptiongroup==1.3.0
flake8==7.1.2
inflection==0.5.1
iniconfig==2.1.0
mccabe==0.7.0
packaging==25.0
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.5
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.13.2
uritemplate==4.1.1
